refactor: move model to different package
