+++ pkg/model/help.go
@@ -1,0 +1,39 @@
+package model
+
+import (
+	"fmt"
+	"reflect"
+	"strings"
+
+	"github.com/charmbracelet/bubbles/key"
+	"github.com/charmbracelet/lipgloss"
+)
+
+func KeyMapInfo(keyMap KeyMap, style lipgloss.Style) []string {
+	v := reflect.ValueOf(keyMap)
+	fields := reflect.VisibleFields(v.Type())
+
+	keys := make([]string, 0)
+	for i := range fields {
+		k := v.Field(i).Interface().(key.Binding)
+		str := k.Help().Key
+		if lipgloss.Width(str) == 0 {
+			str = strings.Join(k.Keys(), ", ")
+		}
+		keys = append(keys, fmt.Sprintf("%v    ", str))
+	}
+
+	desc := make([]string, 0)
+	for i := range fields {
+		k := v.Field(i).Interface().(key.Binding)
+		desc = append(desc, fmt.Sprintf("%v", k.Help().Desc))
+	}
+
+	content := lipgloss.JoinHorizontal(
+		lipgloss.Top,
+		strings.Join(keys, "\n"),
+		strings.Join(desc, "\n"),
+	)
+
+	return strings.Split(style.Render(content), "\n")
+}
