+++ pkg/model/keymap.go
@@ -1,4 +1,4 @@
-package main
+package model
 
 import "github.com/charmbracelet/bubbles/key"
